Validate an AIO configuration file for osad kilo
The schema was initially written from the
openstack_user_config.yml.example documentation which claims that
is_ssh_address and is_container_address are both required. The example
AIO file seems to indicate that they are not (since this was taken from
a working AIO). I've commented those for now until I can confirm with
someone else that these are in fact required.